% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply filters to a response matrix}
\usage{
apply_filters(Y, filters = NULL)
}
\arguments{
\item{Y}{A data.frame with N rows.}

\item{filters}{A list of filters to apply to the response data.}
}
\value{
A data.frame containing only the rows of \code{Y} that pass all filters.
}
\description{
Apply a list of filters to a response matrix \code{Y}. Each filter is a list
with one of two designs. Design 1 applies simple logical comparison based
on the following fields:
\itemize{
\item \code{column}: A string specifying the column name in \code{Y} to filter.
\item \code{fun}: A function to apply to the column values.
\item \code{args}: A list of additional arguments to pass to \code{fun}.
Applying \code{fun} to \code{column} with \code{args} should return a logical, which
is used to filter the data (TRUE to keep, FALSE to discard).
Design 2 filters based on counts of the levels in a factor variable,
based on the following fields:
\item \code{factor}: A string specifying the column name in \code{Y} to use as
a grouping factor
\item \code{min}: An integer specifying the minimum number of observations in
a group (groups with fewer observations than this are discarded)
}
}
