% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{cli}
\alias{cli}
\title{Command-line utility for running CDR-GAM models}
\usage{
cli()
}
\description{
A command-line utility for fitting, evaluating, and plotting CDR-GAM models.
Wraps the \code{main()} function for use in the command line as follows:
\verb{Rscript -e "cdrgam::cli()" <cfg> <model_name> --<option1> <value1> --<option2> <value2> ...}
where \verb{<cfg>} is the path to a YAML configuration file, \verb{<model_name>} is the
name of the model to fit, and \verb{<option1>}, \verb{<option2>}, etc., are any optional
parameters to \code{main()}. Note that, due to limitations of command line parsing in R,
the \code{--eval_partition} option can accept lists, but only as comma-separated (rather
than space-separated) strings.
}
