% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a CDR-GAM model on a dataset}
\usage{
evaluate(model, X, y)
}
\arguments{
\item{model}{A CDR-GAM object (i.e., an \code{mgcv} model fitted with a CDR-GAM
model structure on CDR-GAM-structured data).}

\item{X}{A list containing matrix-valued variables, including a 'mask'
variable indicating whether each observation is valid (1) or invalid (0).
Each other variable should be a matrix of the same dimensionality as
'mask'. This list is typically the output of \code{get_cdr_data()}.}

\item{y}{A numeric vector of response values. Must be the same length as
the number of rows in the matrices in \code{X}.}
}
\value{
A numeric vector representing the log-likelihood of each datapoint
given the model. Dataset likelihood can be computed by summing this vector.
}
\description{
Evaluate a CDR-GAM model on a dataset by computing the log-likelihood of the
data given the model. The model must be a GAM object from the \code{mgcv} package
fitted using one of the following families: "gaussian", "gaulss", or "shash".
}
