% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evaluate_cdrgam}
\alias{evaluate_cdrgam}
\title{Evaluate a CDR-GAM model}
\usage{
evaluate_cdrgam(cfg, model_name, eval_partition = "val", extra_cols = FALSE)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to evaluate}

\item{eval_partition}{A string with the name of the partition on which to
evaluate the model (e.g., 'val', 'test')}

\item{extra_cols}{A logical, whether to include all columns from \code{Y} in
the output}
}
\description{
Evaluate a CDR-GAM model based on a configuration file. The function will
load the model, compute the log-likelihood of the data given the model, and
save the results to disk.
}
