% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_cdrnn}
\alias{fit_cdrnn}
\title{Fit a CDR-GAM model}
\usage{
fit_cdrnn(cfg, model_name, overwrite = FALSE)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to fit}

\item{overwrite}{A logical, whether to overwrite an existing model}
}
\description{
Fit a CDR-GAM model based on a configuration file. The function will load
the data, fit the model, and save the model to disk. If the model already
exists, the function will skip fitting unless the \code{overwrite} flag is set.
}
