% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_cdr_population_code}
\alias{get_cdr_population_code}
\title{Get the CDR-GAM population code}
\usage{
get_cdr_population_code()
}
\value{
A string
}
\description{
Get the string code for the reserved value (representing the population
level)  of CDR-GAM random effects ("!!!POPULATION!!!"). Use for
prediction and plotting of fixed effects.
}
