% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_cdr_quantiles}
\alias{get_cdr_quantiles}
\title{Get vectors of quantiles for numeric columns in a CDR dataset}
\usage{
get_cdr_quantiles(x, n = 101)
}
\arguments{
\item{x}{A list containing matrix-valued variables, including a 'mask'
variable indicating whether each observation is valid (1) or invalid (0).
Each other variable should be a matrix of the same dimensionality as
'mask'. This list is typically the output of \code{get_cdr_data()}.}

\item{n}{An integer specifying the number of quantiles to compute.}
}
\value{
A list containing the quantile vectors of each numeric predictor
in \code{x}.
}
\description{
Compute quantile vectors of numeric predictors in a CDR dataset, excluding
values marked as invalid by the 'mask' column.
}
