% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_formula_from_config}
\alias{get_formula_from_config}
\title{Get a CDR-GAM formula from a configuration}
\usage{
get_formula_from_config(formula_cfg, response_name)
}
\arguments{
\item{formula_cfg}{A formula configuration object}

\item{response_name}{A string with the name of the response variable}
}
\value{
A list of formulas
}
\description{
Get a CDR-GAM formula from a formula configuration object. The
function will return a list of formulas, one for each response parameter.
}
