% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_formula_string}
\alias{get_formula_string}
\title{Compute a standard CDR-GAM formula string}
\usage{
get_formula_string(
  preds,
  k = 5,
  k_t = 10,
  bs = "cr",
  bs_t = "cr",
  ran_gfs = NULL,
  random_intercept = TRUE,
  random_rate = TRUE,
  random_IRFs = TRUE
)
}
\arguments{
\item{preds}{A character vector of predictor names}

\item{k}{A numeric or list of numerics, the degree of the IRF splines for
each predictor. If a single numeric, the same value is used for all
predictors.}

\item{k_t}{A numeric or list of numerics, the degree of the IRF splines for
the time delta variable of each IRF. If a single numeric, the same value
is used for all predictors.}

\item{bs}{A character or list of characters, the basis function for the
IRF splines for each predictor. If a single character, the same value is
used for all IRFs. See \code{?mgcv::smooth.terms} for details.}

\item{bs_t}{A character or list of characters, the basis function for the
IRF splines for the time delta variable of each IRF. If a single character,
the same value is used for all IRFs. See \code{?mgcv::smooth.terms} for details.}

\item{ran_gfs}{A character vector of random grouping factor names. If
provided, the formula will include terms for the random grouping factors.
If NULL, no random effects terms are included.}

\item{random_intercept}{A logical, whether to include random intercept terms
for the random grouping factors. Ignored if \code{ran_gfs} is NULL.}

\item{random_rate}{A logical, whether to include random rate (deconvolutional
intercept) terms for the random grouping factors. Ignored if \code{ran_gfs} is
NULL.}

\item{random_IRFs}{A logical, whether to include random IRF terms for the
random grouping factors. Ignored if \code{ran_gfs} is NULL.}
}
\value{
A string representing the RHS of the model formula
}
\description{
Compute a string representation of the right-hand side (RHS) of a standard
CDR-GAM formula from a set of preditors and optional parameters. The
model will contain a "rate" term (deconvolutional intercepts), additive
impulse response functions (IRFs) for each predictor individually, and
intercept, rate, and IRF terms for each random grouping factor, if
applicable. Model formulas can of course be written by the modeler,
so this is simply a convenience function for typical cases, which
guarantees that the resulting model is a valid CDR-GAM. More complex
terms (e.g., terms involving IRFs of predictor interactions) currently
must be added to the output by hand using string concatenation
(\code{paste()} function).
}
