% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_irf_metadata}
\alias{get_irf_metadata}
\title{Get impulse response function metadata for plots}
\usage{
get_irf_metadata(
  model,
  means,
  sds,
  xlim = c(0, 1),
  irf_name_map = NULL,
  add_rate = TRUE,
  exclude = c("t_delta", "mask")
)
}
\arguments{
\item{model}{A CDR-GAM object (i.e., an \code{mgcv} model fitted with a CDR-GAM
model structure on CDR-GAM-structured data).}

\item{means}{A list of means for each variable in the model. This is
typically the output of \code{get_cdr_means()}.}

\item{sds}{A list of standard deviations for each variable in the model.
This is typically the output of \code{get_cdr_sds()}.}

\item{xlim}{A numeric vector of length 2 specifying the minimum and maximum
values for the x-axis. If NULL, the function will default to 0 and 1.}

\item{irf_name_map}{A list of mappings from IRF names to human-readable
names. This is useful for renaming IRFs that are not easily interpretable
from the default names generated by \code{mgcv} (i.e. most of them).}

\item{add_rate}{A logical value indicating whether to add a rate term to the
IRF list. This is useful for models that include a rate term.}

\item{exclude}{A character vector of term names to exclude from the IRF list.}
}
\value{
A list of IRF metadata, each containing the following elements:
\itemize{
\item irf_name: The name of the IRF
\item term_name: The name of the term (predictor) in the model
\item smooth_name: The name of the smooth term in the model
\item ref: A list of reference values for the IRF
\item xlim: A numeric vector of length 2 specifying the minimum and maximum
values for the x-axis
}
}
\description{
Compute metadata about each impulse response function (IRF) in a CDR-GAM
model. This function is useful because IRFs are implicit in the
specification of \code{mgcv} smooths and are not directly accessible from the
model itself, although they can be inferred from standard \code{mgcv} model
metadata. This function performs this inference and is used by
downstream functions (e.g., high-level plotting functions like
\code{plot_irfs()}).
}
