% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_plot_cfg}
\alias{get_plot_cfg}
\title{Get CDR-GAM plot configuration object from YAML file}
\usage{
get_plot_cfg(path)
}
\arguments{
\item{path}{A string with the path to the YAML file}
}
\value{
A list with the configuration
}
\description{
Get configuration object from YAML file. This function reads a YAML file
and returns a list with the configuration. The configuration is populated
as needed against a set of defaults.
}
