% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Compute CDR-GAM plot data}
\usage{
get_plot_data(
  model,
  smooth_name,
  xvar,
  yvar = NULL,
  X_ref = NULL,
  xaxis = NULL,
  xlim = NULL,
  xres = NULL,
  yaxis = NULL,
  ylim = NULL,
  yres = NULL,
  reshape_3d = FALSE,
  mask_col = "mask"
)
}
\arguments{
\item{model}{A CDR-GAM object (i.e., an \code{mgcv} model fitted with a CDR-GAM
model structure on CDR-GAM-structured data).}

\item{smooth_name}{The name of the smooth term to plot.}

\item{xvar}{The name of the variable to manipulate along the x-axis.}

\item{yvar}{The name of the variable to manipulate along the y-axis. If NULL,
the function will return a line plot (1D manipulation). If provided,
the function will return a surface plot (2D manipulation).}

\item{X_ref}{A list of reference values for some subset of variables in the
model. If NULL, or if a variable is missing, the function will default to
a value of 0 for numeric variables, the first level for factors that define
fixed effects, and the value of \code{cdr_population_code()} for factors that
define random grouping factors.}

\item{xaxis}{A numeric vector of values to use for the x-axis. If NULL, the
function will construct the axis from the value of \code{xlim}.}

\item{xlim}{A numeric vector of length 2 specifying the minimum and maximum
values for the x-axis. If NULL, the function will default to 0 and 1.}

\item{xres}{The number of points to use for the x-axis. If NULL, the function
will default to 1024 for line plots and 32 for surface plots.}

\item{yaxis}{A numeric vector of values to use for the y-axis. If NULL, the
function will construct the axis from the value of \code{ylim}.}

\item{ylim}{A numeric vector of length 2 specifying the minimum and maximum
values for the y-axis. If NULL, the function will default to 0 and 1.}

\item{yres}{The number of points to use for the y-axis. If NULL, the function
will default to 32.}

\item{reshape_3d}{A logical value indicating whether to reshape the output
data into the matrix format expected by \code{plotly}'s 3D plotting
functions. If FALSE, all values will be returned as flattened vectors,
which can be used to e.g., construct heatmaps using \code{ggplot2}.}

\item{mask_col}{A string specifying the name of the column
containing the mask over valid timepoints.}
}
\value{
A list containing the following elements:
\itemize{
\item smooth_name: The name of the smooth term.
\item xvar: The name of the variable manipulated along the x-axis.
\item yvar: The name of the variable manipulated along the y-axis if appicable
(otherwise NULL)
\item x: A numeric vector of x-axis values.
\item y: A numeric vector of y-axis values, if appicable (otherwise NULL)
\item resp: A numeric vector of response values.
\item resp.se: A numeric vector of standard errors for the response values.
}
}
\description{
Compute data (numeric arrays representing lines or surfaces) from a CDR-GAM
model by predicting different terms under systematic manipulations of the
input. This function is very flexible at the expense of user-friendliness.
It can be called directly to create arbitrary plots, but it is typically
called by higher-level functions for generating commonly-needed plot types
(e.g., \code{plot_irfs()}).
}
