% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_smooth_names}
\alias{get_smooth_names}
\title{Get the names of the smooth terms in a GAM object}
\usage{
get_smooth_names(object)
}
\arguments{
\item{object}{A GAM object}
}
\value{
A character vector
}
\description{
Get the names of the smooth terms in a GAM object
}
