% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_time_windows}
\alias{get_time_windows}
\title{Get time windows for each observation in Y}
\usage{
get_time_windows(
  X,
  Y,
  series_ids = NULL,
  forward = FALSE,
  window_length = 32,
  t_delta_cutoff = NULL,
  time_col = "time",
  verbose = TRUE
)
}
\arguments{
\item{X}{A data.frame with M rows. Must contain a column matching \code{time_col}
representing timestamps.}

\item{Y}{A data.frame with N rows. Must contain a column matching \code{time_col}
representing timestamps.}

\item{series_ids}{A character vector of column names in \code{X} and \code{Y} that
identify series groups (a series group is a sequence of observations
that forms a contiguous time series). If NULL, all rows are treated
as a single series.}

\item{forward}{A logical indicating whether to compute the window
forward in time (TRUE) or backward in time (FALSE).}

\item{window_length}{An integer specifying the maximum number of rows
in X to include in each window.}

\item{t_delta_cutoff}{An integer specifying the maximum time difference
between the start (if \code{forward=TRUE}) or end (if \code{forward=FALSE}) of
a window and the timestamp of the corresponding row in \code{Y}. If NULL,
no cutoff is applied.}

\item{time_col}{A string specifying the column name in \code{X} and \code{Y} that
contains timestamps.}

\item{verbose}{A logical indicating whether to print progress messages.}
}
\value{
A list with two elements: 'first_obs' and 'last_obs'. Each
is a vector with N elements, where N is the number of rows in \code{Y}.
'first_obs' contains the row index in \code{X} of the first observation in
the window for each row in \code{Y}. 'last_obs' contains the row index in \code{X}
of the last observation in the window for each row in \code{Y}.
}
\description{
Compute local time windows (row start and end indices in impulse
matrix \code{X}) for each row in response matrix \code{Y}, given a specific window
size and direction. Used to constructed the time-lagged impulse matrix
need to fit a CDR-GAM.
}
