% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.string}
\alias{is.string}
\title{Check whether a variable is a single string}
\usage{
is.string(x)
}
\arguments{
\item{x}{A variable}
}
\value{
\code{TRUE} if \code{x} is a single string, \code{FALSE} otherwise
}
\description{
Check whether a variable is a single string (i.e., a character vector
of length 1)
}
