% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load.cdrgam}
\alias{load.cdrgam}
\title{Load a CDR-GAM object from a file}
\usage{
load.cdrgam(file, ...)
}
\arguments{
\item{file}{A file path}

\item{...}{Additional arguments to \code{readRDS()}}
}
\value{
A CDR-GAM object, as saved by \code{save.cdrgam()},
with the mgcv model object stored in the \code{m} field
and additional metadata stored in other fields.
}
\description{
Load a CDR-GAM object from a file
}
