% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{main}
\alias{main}
\title{High-level CDR-GAM wrapper for fitting, evaluating, and plotting}
\usage{
main(
  cfg,
  model_name,
  fit = TRUE,
  eval_partition = c("train", "val"),
  plot = TRUE,
  plot_cfg = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to fit}

\item{fit}{A logical, whether to fit the model}

\item{eval_partition}{A string or a list of strings with the names of the
partition(s) (e.g., train, val, test) on which to evaluate}

\item{plot}{A logical, whether to visualize the model estimates}

\item{plot_cfg}{A configuration object or a string with the path to a YAML
file with additional plot configuration settings (useful for overriding
defaults used in the main config)}

\item{overwrite}{A logical, whether to refit and overwrite an existing
model file}
}
\description{
A convenience function for fitting, evaluating, and plotting a CDR-GAM model
using a configuration file. The function will fit the model if \code{fit} is
\code{TRUE}, evaluate the model if \code{eval_partition} is provided, and visualize
the model if \code{plot} is \code{TRUE}.
}
