% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{make_jobs}
\alias{make_jobs}
\title{Command-line utility for making slurm jobs from CDR-GAM config files}
\usage{
make_jobs()
}
\description{
A command-line utility for generating slurm job scripts from CDR-GAM config
files. The utility generates one job script per model in each config file.
The job scripts are written to the current working directory.
}
