% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation test a difference between models}
\usage{
permutation_test(
  a0,
  a1,
  n_iter = 10000,
  n_tails = 2,
  agg = "sum",
  statistic = "logLik",
  nested = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{a0}{A numeric vector of the base model's values.}

\item{a1}{A numeric vector of the alternative model's values.}

\item{n_iter}{The number of iterations to perform.}

\item{n_tails}{The number of tails to use in the test.}

\item{agg}{The aggregation function to use. Either 'sum' or 'mean'.}

\item{statistic}{The statistic to use. Either 'logLik' or 'error'.}

\item{nested}{A logical value indicating if the test is nested. If TRUE, the
test is only performed if the observed difference is positive (i.e., if
the alternative model is better than the base model). Otherwise, the test
is always performed.}

\item{verbose}{A logical value indicating if progress should be printed.}
}
\value{
A list with the p-value, the observed difference, and the samples.
}
\description{
This function performs a permutation test to determine if the difference
between two models is significant.
}
