% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_cdrgam}
\alias{plot_cdrgam}
\title{Plot a CDR-GAM model}
\usage{
plot_cdrgam(cfg, model_name, plot_cfg = NULL, dump_plot_data = FALSE)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to plot}

\item{plot_cfg}{A configuration object or a string with the path to a YAML
file with additional plot configuration settings (useful for overriding
defaults used in the main config)}

\item{dump_plot_data}{A logical, whether to save the data used to generate the
plots to disk as a CSV file. The saved data will be in the same location
as the plot, and have the same basename (just a different file
extension). The data will be in long format, with columns for the x-axis,
y-axis, and any other relevant information, with different subplots
distinguished as needed by the \code{color} column. This can be used to
reload the data into other software environments in order to customize
the look and contents of plots.}
}
\description{
Visualize the estimates of a CDR-GAM model based on a configuration file.
The function will load the model, produce plots, and save the plots to disk.
}
