% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_cdrnn}
\alias{plot_cdrnn}
\title{Plot a CDR-GAM model}
\usage{
plot_cdrnn(cfg, model_name, plot_cfg = NULL)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to plot}

\item{plot_cfg}{A configuration object or a string with the path to a YAML
file with additional plot configuration settings (useful for overriding
defaults used in the main config)}
}
\description{
Visualize the estimates of a CDR-GAM model based on a configuration file.
The function will load the model, produce plots, and save the plots to disk.
}
