% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_curvature}
\alias{plot_curvature}
\title{Generate curvature plots}
\usage{
plot_curvature(
  model,
  means,
  sds,
  quantiles = NULL,
  range = 1,
  xlim = NULL,
  t_delta_ref = 0,
  irf_name_map = NULL,
  xlabel = "Predictor",
  ylabel = "y",
  legend = TRUE,
  exclude = c("t_delta", "mask")
)
}
\arguments{
\item{model}{A CDR-GAM object (i.e., an \code{mgcv} model fitted with a CDR-GAM
model structure on CDR-GAM-structured data).}

\item{means}{A list of means for each variable in the model. This is
typically the output of \code{get_cdr_means()}.}

\item{sds}{A list of standard deviations for each variable in the model.
This is typically the output of \code{get_cdr_sds()}.}

\item{quantiles}{A list of quantiles for each variable in the model. This is
typically the output of \code{get_cdr_quantiles()}.}

\item{range}{A numeric value specifying the range of quantiles to use for
the x-axis. For example, a value of 0.9 would use the 5th to 95th
quantiles for the x-axis. Overridden by \code{xlim}.}

\item{xlim}{A list of numeric vectors of length 2 specifying the minimum and
maximum values for the x-axis for each term in the model. If NULL, the
function will default to 0 and 1.}

\item{t_delta_ref}{A numeric value specifying the delay (t_delta) value at
which to evaluate the curvature.}

\item{irf_name_map}{A list of mappings from IRF names to human-readable
names. This is useful for renaming IRFs that are not easily interpretable
from the default names generated by \code{mgcv} (i.e. most of them).}

\item{xlabel}{A string specifying the x-axis label}

\item{ylabel}{A string specifying the y-axis label}

\item{legend}{A logical value indicating whether to include a legend}

\item{exclude}{A character vector of term names to exclude from the IRF list.}
}
\value{
A ggplot2 object
}
\description{
Generate plots of the curvature of the response surface of a CDR-GAM model
at a specific delay (t_delta) value. This function is a high-level wrapper
around \code{get_irf_metadata()} and \code{get_plot_data()}.
}
