% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_irfs}
\alias{plot_irfs}
\title{Generate impulse response function line plots}
\usage{
plot_irfs(
  model,
  response_param = 1,
  means = NULL,
  sds = NULL,
  xlim = c(0, 1),
  irf_name_map = NULL,
  xlabel = "Delay (s)",
  ylabel = "y",
  legend = TRUE,
  exclude = c("t_delta", "mask"),
  t_delta_col = "t_delta",
  mask_col = "mask"
)
}
\arguments{
\item{model}{A CDR-GAM object (i.e., an \code{mgcv} model fitted with a CDR-GAM
model structure on CDR-GAM-structured data).}

\item{response_param}{An integer specifying the index of the response
parameter to use for the IRF. E.g., if the response is Gaussian,
\code{response_param=1} refers to the mean parameter, while \code{response_param=2}
refers to the standard deviation parameter.}

\item{means}{A list of means for each variable in the model. This is
typically the output of \code{get_cdr_means()}. If NULL, the function will
default to 0 for all numeric variables and the first level for all
factor variables.}

\item{sds}{A list of standard deviations for each variable in the model.
This is typically the output of \code{get_cdr_sds()}. If NULL, the function
will default to 1.}

\item{xlim}{A numeric vector of length 2 specifying the minimum and maximum
values for the x-axis. If NULL, the function will default to 0 and 1.}

\item{irf_name_map}{A list of mappings from IRF names to human-readable
names. This is useful for renaming IRFs that are not easily interpretable
from the default names generated by \code{mgcv} (i.e. most of them).}

\item{xlabel}{A string specifying the x-axis label}

\item{ylabel}{A string specifying the y-axis label}

\item{legend}{A logical value indicating whether to include a legend}

\item{exclude}{A character vector of term names to exclude from the IRF list.}

\item{t_delta_col}{A string specifying the name of the column
containing the difference in time between impulses and response.}

\item{mask_col}{A string specifying the name of the column
containing the mask over valid timepoints.}
}
\value{
A ggplot2 object
}
\description{
Generate line plots of impulse response functions (IRFs) from a CDR-GAM
model. This function is a high-level wrapper around \code{get_irf_metadata()}
and \code{get_plot_data()}.
}
