% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lines}
\alias{plot_lines}
\title{Plot lines}
\usage{
plot_lines(
  data_2d,
  xlabel = "x",
  ylabel = "y",
  legend = TRUE,
  add_origin = TRUE
)
}
\arguments{
\item{data_2d}{A list of lists, each containing the following
keys: "x", "resp", and (optionally) "resp.se". This is typically
the output of \code{get_plot_data()}.}

\item{xlabel}{A string specifying the x-axis label}

\item{ylabel}{A string specifying the y-axis label}

\item{legend}{A logical value indicating whether to include a legend}

\item{add_origin}{A logical value indicating whether to add lines
at x=0 and y=0}
}
\value{
A ggplot2 object
}
\description{
Plot lines with confidence intervals. Each line is represented by a
numeric vector of x-axis values, a numeric vector of y-axis values
(called "resp"), and a numeric vector of standard errors for the
y-axis values (called "resp.se").
}
