% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report}
\alias{report}
\title{Report message to the console}
\usage{
report(x)
}
\arguments{
\item{x}{A string}
}
\description{
Write a string to standard error. Useful alternative to \code{message()}
that does not append a newline, allowing it to be used for progress
reporting.
}
