% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{reshape_3d_plot}
\alias{reshape_3d_plot}
\title{Reshape 3D plot data}
\usage{
reshape_3d_plot(x, y, z, z.se = NULL, nx = NULL)
}
\arguments{
\item{x}{A numeric vector of x-axis values}

\item{y}{A numeric vector of y-axis values}

\item{z}{A numeric vector of response values}

\item{z.se}{A numeric vector of standard errors for the response values}

\item{nx}{An integer specifying the number of x-axis values}
}
\value{
A list containing reshaped 2D matrices for for all inputs
}
\description{
Reshape flat vector-valued 3D plot data into a list of 2D matrices for
plotting with \code{plotly}. Vector lengths must either be perfect squares
(if \code{nx} is not provided) or must be divisible by \code{nx}.
}
