% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save.cdrgam}
\alias{save.cdrgam}
\title{Save a CDR-GAM object to a file}
\usage{
save.cdrgam(obj, file, clean = TRUE, keep_model = FALSE, ...)
}
\arguments{
\item{obj}{A CDR-GAM object, with the mgcv model object
stored in the \code{m} field and additional metadata stored
in other fields.}

\item{file}{A file path}

\item{clean}{A logical value indicating whether to clean
the data from the GAM object before saving, which can
dramatically reduce disk and memory consumption at the
loss of some functionality (see \code{clean_data_from_gam()}
for details)}

\item{keep_model}{A logical value indicating whether to keep
the main model matrix in the object. If \code{FALSE}, the
\code{model} field of the GAM object will be removed, resulting
in significant additional savings but breaking some
native \code{mgcv} functionality like plotting or prediction
without supplying a dataset. Ignored if \code{clean=FALSE}. See
\code{clean_data_from_gam()} for details.}

\item{...}{Additional arguments to \code{saveRDS()}}
}
\description{
Save a CDR-GAM object to a file
}
