% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{size_by_element}
\alias{size_by_element}
\title{Print the memory size of all child elements of an object}
\usage{
size_by_element(obj)
}
\arguments{
\item{obj}{An object}
}
\description{
Print the memory size of all child elements of an object.
Useful for debugging memory usage.
}
