% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{test}
\alias{test}
\title{Command-line utility for statistically comparing two CDR-GAM models}
\usage{
test()
}
\description{
A command-line utility for statistically comparing two CDR-GAM models. The
utility wraps the \code{test_cdrgam()} function for use in the command line as
follows:
\verb{Rscript -e "cdrgam::test()" <cfg0> <model0> <model1> --<option1> <value1> --<option2> <value2> ...}
where \verb{<cfg0>} is the path to the first model configuration file, \verb{<model0>}
is the name of the first model, \verb{<model1>} is the name of the second model, and
\verb{<option1>}, \verb{<option2>}, etc., are any optional parameters to \code{test_cdrgam()}.
Note that, due to limitations of command line parsing in R, the \code{--eval_partition}
option can accept lists, but only as comma-separated (rather than space-separated) strings.
}
