% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{test_cdrgam}
\alias{test_cdrgam}
\title{Perform a paired permutation test of difference between two models}
\usage{
test_cdrgam(
  cfg0,
  model_name0,
  model_name1,
  cfg1 = NULL,
  eval_partition = "val",
  statistic = "logLik",
  output_dir = NULL,
  ...
)
}
\arguments{
\item{cfg0}{A configuration object or a string with the path to a YAML file}

\item{model_name0}{A string with the name of the base model}

\item{model_name1}{A string with the name of the alternative model}

\item{cfg1}{A configuration object or a string with the path to a YAML file.
If NULL, \code{cfg0} will be used in its place.}

\item{eval_partition}{A string or a list of strings with the names of the
partition(s) (e.g., train, val, test) on which to evaluate}

\item{statistic}{A string with the name of the statistic to use}

\item{output_dir}{A string with the path to the output directory}

\item{...}{Additional arguments to pass to the \code{permutation_test()} function}
}
\description{
Perform a paired permutation test of the difference between two models
based on a statistic (e.g., log-likelihood) computed on a partition of the
data. The function will load the data, perform the test, and save the
results to disk.
}
