% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_cfg}
\alias{validate_cfg}
\title{Validate a CDR-GAM configuration}
\usage{
validate_cfg(cfg, model_name)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to validate}
}
\description{
Validate a CDR-GAM configuration object to ensure that it contains the
required fields.
}
