% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_cdrgam}
\alias{fit_cdrgam}
\title{Fit a CDR-GAM model}
\usage{
fit_cdrgam(
  cfg,
  model_name,
  overwrite = FALSE,
  clean = TRUE,
  keep_model = FALSE
)
}
\arguments{
\item{cfg}{A configuration object or a string with the path to a YAML file}

\item{model_name}{A string with the name of the model to fit}

\item{overwrite}{A logical, whether to overwrite an existing model}

\item{clean}{A logical value indicating whether to clean
the data from the GAM object before saving, which can
dramatically reduce disk and memory consumption at the
loss of some functionality (see \code{clean_data_from_gam()}
for details)}

\item{keep_model}{A logical value indicating whether to keep
the main model matrix in the object. If \code{FALSE}, the
\code{model} field of the GAM object will be removed, resulting
in significant additional savings but breaking some
native \code{mgcv} functionality like plotting or prediction
without supplying a dataset. Ignored if \code{clean=FALSE}. See
\code{clean_data_from_gam()} for details.}
}
\description{
Fit a CDR-GAM model based on a configuration file. The function will load
the data, fit the model, and save the model to disk. If the model already
exists, the function will skip fitting unless the \code{overwrite} flag is set.
}
