% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_cdr_sds}
\alias{get_cdr_sds}
\title{Get standard deviations of numeric columns in a CDR dataset}
\usage{
get_cdr_sds(x, mask_col = "mask")
}
\arguments{
\item{x}{A list containing matrix-valued variables, including a 'mask'
variable indicating whether each observation is valid (1) or invalid (0).
Each other variable should be a matrix of the same dimensionality as
'mask'. This list is typically the output of \code{get_cdr_data()}.}

\item{mask_col}{A string specifying the name assigned to the column
containing the mask over valid timepoints.}
}
\value{
A list containing the SDs of each numeric predictor in \code{x}.
}
\description{
Compute the SDs of numeric predictors in a CDR dataset, excluding
values marked as invalid by the 'mask' column.
}
