% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_formula_string}
\alias{get_formula_string}
\title{Compute a standard CDR-GAM formula string}
\usage{
get_formula_string(
  irfs = NULL,
  k = 5,
  k_t = 10,
  bs = "cr",
  bs_t = "cr",
  use_intercept = TRUE,
  use_rate = TRUE,
  ran_gf = NULL,
  t_delta_col = "t_delta",
  mask_col = "mask"
)
}
\arguments{
\item{irfs}{A character vector of predictor names, or a list of character
vectors of predictor names, or a list of lists of character vectors of
predictor names. If a list, each element of the list--i.e., each
(possibly singleton) set of predictors--defines a separate smooth
that will be added to the model. If a list of lists, each element must
be a pair (list of length 2) in which the first subelement (named
"inputs") is a character vector of predictors to include in the
inputs to the IRF, and the second subelement (named "impulses") is
a character vector of impulses to convolve using the IRF.
A smooth involving multiple predictors will represent an IRF of the
interaction of all predictors in the set. For example, \code{c('a', 'b')} will
create independent IRFs for \code{a} and \code{b}, whereas
\code{list('a', 'b', c('a', 'b'))} will additionally create an IRF of the
interaction of \code{a} and \code{b}.}

\item{k}{A numeric or list of numerics, the degree of the IRF splines for
each predictor. If a single numeric, the same value is used for all
predictors.}

\item{k_t}{A numeric or list of numerics, the degree of the IRF splines for
the time delta variable of each IRF. If a single numeric, the same value
is used for all predictors.}

\item{bs}{A character or list of characters, the basis function for the
IRF splines for each predictor. If a single character, the same value is
used for all IRFs. See \code{?mgcv::smooth.terms} for details. If NULL,
the IRF will be constrained to be linear in the predictor.}

\item{bs_t}{A character or list of characters, the basis function for the
IRF splines for the time delta variable of each IRF. If a single character,
the same value is used for all IRFs. See \code{?mgcv::smooth.terms} for details.}

\item{use_intercept}{A logical, whether to include an intercept term}

\item{use_rate}{A logical, whether to a rate (deconvolutional
intercept) term}

\item{ran_gf}{A string containing the name of a random grouping factor. If
provided, the formula will interact all terms with the random grouping factor.
If NULL, the string will represent fixed effects terms.}

\item{t_delta_col}{A string specifying the name of the column
containing the difference in time between impulses and response.}

\item{mask_col}{A string specifying the name of the column
containing the mask over valid timepoints.}
}
\value{
A string representing the RHS of the model formula
}
\description{
Compute a string representation of the right-hand side (RHS) of a standard
CDR-GAM formula from a set of preditors (or predictor sets in the case of
interactions) and optional parameters. Model formulas can of course be
hand written, so this is simply a convenience function for typical cases,
which guarantees that the resulting model is a valid CDR-GAM. However,
since the function returns a string, the output can be edited as needed by
the modeler in special cases. The function returns either a set of fixed
effects terms or a set of random effects terms (if \code{ran_gf} is provided).
Thus, to construct mixed models, simply call this function multiple times,
once for the fixed effects specification and once for each random effects
specification, and concatenate the results with a \code{+} separator.
}
