% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_formula_string}
\alias{get_formula_string}
\title{Compute a standard CDR-GAM formula string}
\usage{
get_formula_string(
  preds,
  k = 5,
  k_t = 10,
  bs = "cr",
  bs_t = "cr",
  ran_gfs = NULL,
  random_intercept = TRUE,
  random_rate = TRUE,
  random_IRFs = TRUE
)
}
\arguments{
\item{preds}{A character vector of predictor names, or a list of character
vectors of predictor names. If a list, each element of the list--i.e.,
each (possibly singleton) set of predictors--defines a separate smooth
that will be added to the model. A smooth involving multiple predictors
will represent an IRF of the interaction of all predictors in the set.
For example, \code{c('a', 'b')} will create independent IRFs for \code{a} and \code{b},
whereas \code{list('a', 'b', c('a', 'b'))} will additionally create an IRF
of the interaction of \code{a} and \code{b}.}

\item{k}{A numeric or list of numerics, the degree of the IRF splines for
each predictor. If a single numeric, the same value is used for all
predictors.}

\item{k_t}{A numeric or list of numerics, the degree of the IRF splines for
the time delta variable of each IRF. If a single numeric, the same value
is used for all predictors.}

\item{bs}{A character or list of characters, the basis function for the
IRF splines for each predictor. If a single character, the same value is
used for all IRFs. See \code{?mgcv::smooth.terms} for details. If NULL,
the IRF will be constrained to be linear in the predictor.}

\item{bs_t}{A character or list of characters, the basis function for the
IRF splines for the time delta variable of each IRF. If a single character,
the same value is used for all IRFs. See \code{?mgcv::smooth.terms} for details.}

\item{ran_gfs}{A character vector of random grouping factor names. If
provided, the formula will include terms for the random grouping factors.
If NULL, no random effects terms are included.}

\item{random_intercept}{A logical, whether to include random intercept terms
for the random grouping factors. Ignored if \code{ran_gfs} is NULL.}

\item{random_rate}{A logical, whether to include random rate (deconvolutional
intercept) terms for the random grouping factors. Ignored if \code{ran_gfs} is
NULL.}

\item{random_IRFs}{A logical, whether to include random IRF terms for the
random grouping factors. Ignored if \code{ran_gfs} is NULL.}
}
\value{
A string representing the RHS of the model formula
}
\description{
Compute a string representation of the right-hand side (RHS) of a standard
CDR-GAM formula from a set of preditors (or predictor sets in the case of
interactions) and optional parameters. The model will contain a "rate"
term (deconvolutional intercepts), additive impulse response functions
(IRFs) for each predictor set individually, and corresponding
intercept, rate, and IRF terms for each random grouping factor, if
applicable. Model formulas can of course be written by the modeler,
so this is simply a convenience function for typical cases, which
guarantees that the resulting model is a valid CDR-GAM. However, since
the function returns a string, the output can be edited as needed by
the modeler in special cases.
}
